+++ lib/rubocop/formatter/formatter_set.rb
@@ -36,6 +36,7 @@ module RuboCop
         @comments[file] = options[:comments]
         @excepted_cops = options[:excepted_cops] || []
         @only_cops = options[:only_cops] || []
+        @config_store = options[:config_store]
         each { |f| f.file_started(file, options) }
       end
 
@@ -43,7 +44,8 @@ module RuboCop
         if @cop_disabled_line_ranges[file].any? &&
            # Don't check unneeded disable if --only or --except option is
            # given, because these options override configuration.
-           @excepted_cops.empty? && @only_cops.empty?
+           @excepted_cops.empty? && @only_cops.empty? &&
+           @config_store.for(file)['Lint/UnneededDisable']['Enabled']
           cop = Cop::Lint::UnneededDisable.new
           cop.check(file, offenses, @cop_disabled_line_ranges, @comments)
           offenses += cop.offenses